+++ controllers/index.js
@@ -1,0 +1,3 @@
+const ctrlContacts = require("./contacts.js");
+
+module.exports = { ctrlContacts };
